Organised Project Files and Removed Debug Logs and Added Better Comments
