+++ Assembly-CSharp.csproj
@@ -65,20 +65,20 @@
     <Analyzer Include="/Applications/Unity/Hub/Editor/6000.0.43f1/Unity.app/Contents/Tools/Unity.SourceGenerators/Unity.UIToolkit.SourceGenerator.dll" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Assets/Scripts/Enemy/TerminalUIHandler.cs" />
     <Compile Include="Assets/Scripts/Timer.cs" />
+    <Compile Include="Assets/Scripts/Enemy/SpawnEnemy.cs" />
     <Compile Include="Assets/Scripts/PauseMenu.cs" />
-    <Compile Include="Assets/Scripts/EnemyFollow.cs" />
-    <Compile Include="Assets/Scripts/EnemyHoverIndicator.cs" />
+    <Compile Include="Assets/Scripts/GameMechanics/TerminalUIHandler.cs" />
+    <Compile Include="Assets/Scripts/Enemy/EnemyHoverIndicator.cs" />
     <Compile Include="Assets/Scripts/Bullet.cs" />
-    <Compile Include="Assets/Scripts/EnemyBullet.cs" />
-    <Compile Include="Assets/Scripts/Enemy/SoftDevQuestionManager.cs" />
+    <Compile Include="Assets/Scripts/Enemy/EnemyFollow.cs" />
+    <Compile Include="Assets/Scripts/Enemy/EnemyBullet.cs" />
+    <Compile Include="Assets/Scripts/GameMechanics/SoftDevQuestionManager.cs" />
     <Compile Include="Assets/Scripts/Input/InputManager.cs" />
     <Compile Include="Assets/Scripts/Enemy/EnemyClickHandler.cs" />
     <Compile Include="Assets/Scripts/PlayerMovement.cs" />
     <Compile Include="Assets/Scripts/MusicManager.cs" />
-    <Compile Include="Assets/Scripts/HealthManager.cs" />
-    <Compile Include="Assets/Scripts/SpawnEnemy.cs" />
+    <Compile Include="Assets/Scripts/Enemy/HealthManager.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="Assets/TextMesh Pro/Shaders/TMPro.cginc" />
