+++ Assets/Art/Animations/SDIdleLeft.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7a6bb159b5b1043418a8154dbfb24476
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
