+++ Assets/Art/Animations/SDIdleRight.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 77f86b496bcba2646b95ff9b5764232c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
