+++ Assets/Art/Animations/SDIdleUp.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3935ecc23a374734d8ae85e765af981d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
