--- Assets/Art/SDIdleDown.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e68150798d71f2149bc1ded07a2a8b85
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
