--- Assets/Art/SDIdleLeft.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7a6bb159b5b1043418a8154dbfb24476
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
