--- Assets/Art/SDIdleUp.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3935ecc23a374734d8ae85e765af981d
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
