--- Assets/Scripts/Enemy/SoftDevQuestionManager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: e66145fb173e74884b3c8704351b66c8
