--- Assets/Scripts/Enemy/TerminalUIHandler.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 8675dfbd4204f40b2aecfa2fdd4fac19
