--- Assets/Scripts/EnemyBullet.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 8822de1321836b14db686728a56a5c08
