--- Assets/Scripts/EnemyFollow.cs
@@ -1,15 +1,0 @@
-using UnityEngine;
-
-public class EnemyFollow : MonoBehaviour
-{
-    public Transform player;     
-    public float speed = 2f;      
-
-    void Update()
-    {
-        if (player == null) return;
-
-        Vector3 direction = (player.position - transform.position).normalized;
-        transform.position += direction * speed * Time.deltaTime;
-    }
-}
