--- Assets/Scripts/EnemyFollow.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 958cbd43a4ce7e343a4e3281f48dfff7
