--- Assets/Scripts/EnemyHoverIndicator.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 69edf2563db94417392263827d594dfd
