+++ Assets/Scripts/GameMechanics.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1e7e76b018c9247d28d55d4d7f1c701e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
