+++ Assets/Scripts/GameMechanics/TerminalUIHandler.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 8675dfbd4204f40b2aecfa2fdd4fac19
