--- Assets/Scripts/HealthManager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 614b018ff59f4734aa372d045ffae9bb
