--- Assets/Scripts/SpawnEnemy.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 8258164c36b4ff84c831993b3c398bd1
